fix(swap): use modified cap strategy interface
